ref: Move debug_meta into a separate package
This creates a new `DebugImagesIntegration`, which can be configured
with an optional filter. By default, *all* events will get a list of
loaded libraries attached, depending on OS support.
