+++ sentry-core/Cargo.toml
@@ -19,7 +19,6 @@ all-features = true
 default = ["with_client_implementation"]
 with_client_implementation = ["im", "url", "rand"]
 with_debug_to_log = ["log"]
-with_debug_meta = ["findshlibs", "with_client_implementation"]
 with_test_support = []
 
 [dependencies]
@@ -29,7 +28,6 @@ sentry-types = { path = "../sentry-types", version = "0.15.0" }
 env_logger = { version = "0.7.1", optional = true }
 lazy_static = "1.4.0"
 im = { version = "14.2.0", optional = true }
-findshlibs = { version = "0.7.0", optional = true }
 rand = { version = "0.7.3", optional = true }
 regex = { version = "1.3.4", optional = true }
