+++ sentry-core/src/lib.rs
@@ -81,10 +81,6 @@
 //!
 //! * `with_client_implementation`: Turns on the real client implementation.
 //!
-//! Additional context:
-//!
-//! * `with_debug_meta`: Adds debug meta to reported events (permits server side symbolication).
-//!
 //! Additional integrations:
 //!
 //! * `with_debug_to_log`: When enabled sentry will debug log to a debug log at all times.
