+++ sentry/Cargo.toml
@@ -26,7 +26,6 @@ with_backtrace = ["sentry-backtrace"]
 with_panic = ["sentry-panic"]
 with_failure = ["sentry-failure"]
 with_debug_to_log = ["log", "sentry-core/with_debug_to_log"]
-with_debug_meta = ["sentry-core/with_debug_meta"]
 with_test_support = ["sentry-core/with_test_support"]
 with_rustls = ["reqwest/rustls-tls"]
 with_native_tls = ["reqwest/default-tls"]
